requirements.txt
requests>=2.32.3
python-dotenv>=1.0.1
urllib3>=2.2.2
pytest>=8.3.3
